Server Functions/memorylib-video-pipeline/requirements.txt
# Video Memory Pipeline - Dependencies

# Gemini API (via AiHubMix)
google-genai>=0.3.0

# Environment variables
python-dotenv>=1.0.1
